+++ app/presenters/conversations/event_data_presenter.rb
@@ -53,8 +53,8 @@ class Conversations::EventDataPresenter < SimpleDelegator
     {
       agent_last_seen_at: agent_last_seen_at.to_i,
       contact_last_seen_at: contact_last_seen_at.to_i,
-      last_activity_at: last_activity_at.to_i,
-      timestamp: last_activity_at.to_i,
+      last_activity_at: last_activity_at.to_f,
+      timestamp: last_activity_at.to_f,
       created_at: created_at.to_i,
       updated_at: updated_at.to_f
     }
